[Fix] post api 예외처리 추가

+++ src/posts/posts.service.ts
@@ -1,7 +1,12 @@
-import { Injectable } from '@nestjs/common';
+import {
+  Injectable,
+  NotFoundException,
+  ForbiddenException,
+} from '@nestjs/common';
 import { PrismaService } from 'src/prisma/prisma.service';
 import { CreatePostDto } from './dto/create-post.dto';
 import { PostParamDto } from './dto/postParam.dto';
+
 @Injectable()
 export class PostsService {
   constructor(private prisma: PrismaService) {}
@@ -19,9 +24,15 @@ export class PostsService {
   }
   //id기반 포스트 검색
   async getPost(dto: PostParamDto, userId: string) {
-    return this.prisma.post.findFirst({
+    const post = await this.prisma.post.findFirst({
       where: { id: dto.postId, userId: userId },
     });
+
+    if (!post) {
+      throw new NotFoundException('해당 게시글을 찾을 수 없습니다.');
+    }
+
+    return post;
   }
   // 전체 포스트 조회
   async getPosts(userId: string) {
@@ -30,9 +41,21 @@ export class PostsService {
     });
   }
   // 단일 포스트 삭제
-  async deletePost(dto: PostParamDto) {
+  async deletePost(dto: PostParamDto, userId: string) {
+    const post = await this.prisma.post.findUnique({
+      where: { id: dto.postId },
+    });
+
+    if (!post) {
+      throw new NotFoundException('해당 게시글이 존재하지 않습니다.');
+    }
+
+    if (post.userId !== userId) {
+      throw new ForbiddenException('삭제 권한이 없습니다.');
+    }
+
     return this.prisma.post.delete({
-      where: { id: dto.postId }, // delete는 복합유니크조건이 아니면 단일 조건만 가능함.
+      where: { id: dto.postId },
     });
   }
 }
